glomap-preprocess/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
google-cloud-logging